pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pytest-appium-scheduler"
version = "0.1.0"
description = "Pytest plugin for Appium device scheduling and driver lifecycle management."
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
dependencies = [
  "pytest>=7.0",
  "pytest-xdist>=3.0",
  "PyYAML>=6.0",
  "appium-python-client>=5.0",
]
classifiers = [
  "Framework :: Pytest",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Testing",
]
keywords = ["pytest", "appium", "xdist", "scheduler", "mobile-testing"]

[project.entry-points.pytest11]
appium_scheduler = "pytest_appium_scheduler.plugin"

[tool.setuptools.packages.find]
include = ["pytest_appium_scheduler*"]

[tool.setuptools]
license-files = ["LICENSE"]
